requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
